extract 404.js into its own file & fix security for 404.html
putting 404.js into its own file lets us remove the squirrelly
escaping in our regexp code.

the hack to make security work correctly on 404 pages is to call
wiki.security(user) after the /client.js loads.

we're doing a lot of hacking to make lineup urls work in static
sites.

This seems to be the order of events that matter in the
bootstrap life cycle of the wiki client:

1) ensure there are <div class="page"> for each page in the lineup
2) load /client.js
3) assign security variables: isAuthenticated, isClaimed, isOwner,
   ownerName, seedNeighbors, and user
4) call wiki.security(user)

step 1) ensures that in step 2) wiki client correctly loads the page
json for each page in the lineup. Step 3) ensures that we are
correctly logged out and that we know the ownerName when step 4 is
called.

So although this code is complicated, and hard to follow, and will be
hard to change, it will also work correctly without introducing
breaking changes to the wiki-client code.